discord_bot/requirements.txt
discord.py==1.7.3
discord_py_slash_command==3.0.3
fuzzywuzzy==0.18.0
matplotlib==3.3.3
pandas==1.1.3
requests==2.25.1
